Implementation-phase-2/backend/requirements.txt
Flask==2.3.2
google-cloud-bigquery==3.11.0
requests==2.31.0
Jinja2==3.1.2
gunicorn==20.1.0
python-dotenv==1.0.0
Werkzeug==2.3.5
